+++ bin/ParseDb.py
@@ -2,37 +2,35 @@
 """
 Parses tab delimited database files
 """
-
-__author__    = 'Jason Anthony Vander Heiden'
-__copyright__ = 'Copyright 2014 Kleinstein Lab, Yale University. All rights reserved.'
-__license__   = 'Creative Commons Attribution-NonCommercial-ShareAlike 3.0 Unported'
-__version__   = '0.2.3'
-__date__      = '2015.07.22'
+# Info
+__author__ = 'Jason Anthony Vander Heiden'
+from changeo import __version__, __date__
 
 # Imports
-import csv, os, re, sys, textwrap
+import os
+import re
 from argparse import ArgumentParser
-from itertools import izip
 from collections import OrderedDict
+from itertools import izip
+from textwrap import dedent
 from time import time
 from Bio import SeqIO
 from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
 from Bio.Alphabet import IUPAC
 
-# IgCore imports
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
-from IgCore import default_delimiter, default_out_args
-from IgCore import flattenAnnotation 
-from IgCore import getOutputHandle, printLog, printProgress, printMessage
-from IgCore import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
-from DbCore import countDbFile, readDbFile, getDbWriter
+# Presto and changeo imports
+from presto.Defaults import default_delimiter, default_out_args
+from presto.Annotation import flattenAnnotation
+from presto.Commandline import CommonHelpFormatter, getCommonArgParser, parseCommonArgs
+from presto.IO import getOutputHandle, printLog, printProgress, printMessage
+from changeo.IO import getDbWriter, readDbFile, countDbFile
 
 # Defaults
 default_id_field = 'SEQUENCE_ID'
 default_seq_field = 'SEQUENCE_IMGT'
 default_germ_field = 'GERMLINE_IMGT_D_MASK'
-default_index_field = 'INDEX'
+default_index_field = "INDEX"
 
 # TODO:  convert SQL-ish operations to modify_func() as per ParseHeaders
 
@@ -874,7 +872,7 @@ def getArgParser():
     an ArgumentParser object
     """
     # Define input and output field help message
-    fields = textwrap.dedent(
+    fields = dedent(
              '''
              output files:
                sequences      FASTA formatted sequences output from the subcommands fasta